deployment/bed-classifier/pyproject.toml
[tool.poetry]
name = "bed-classifier"
version = "0.1.0"
description = ""
authors = ["OshkarVTec <a01275287@tec.mx>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.5.1"
torchvision = "^0.20.1"
pillow = "^11.0.0"
datetime = "^5.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
